Azure_AI/semantic_kernel/multi_agents/requirements.txt
azure-identity
agent-framework
python-dotenv